Move NonCopyable, DoesThrow and DoesNotThrow to TestHelpers.hpp
